requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
imaplib2==3.6
email-validator==2.1.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.7
aiofiles==23.2.1
openai==0.28.1
schedule==1.2.1
python-dateutil==2.8.2
PyMuPDF==1.23.8
Pillow==10.0.0
beautifulsoup4==4.12.2
typing-extensions>=4.8.0
pdfminer.six==20221105
PyPDF2==3.0.1
pytesseract==0.3.10
pdfplumber==0.10.3
psutil==5.9.6
pytz==2023.3
pymongo==4.6.0
motor==3.3.2